+++ sdk/lib/_internal/js_runtime/lib/io_patch.dart
@@ -459,6 +459,11 @@ class SecurityContext {
   static SecurityContext get defaultContext {
     throw new UnsupportedError("default SecurityContext getter");
   }
+
+  @patch
+  static bool get alpnSupported {
+    throw new UnsupportedError("SecurityContext alpnSupported getter");
+  }
 }
 
 @patch